Add modularization for att_patch and parse arguments
- Modularize the code by separating responsible functionalities into different files: `args_parser.rs`, `ast_interaction.rs`, `chat_interaction.rs`, and `diff_formats.rs`.
- Implement argument parsing in `args_parser.rs`.
- Relocate code related to AST interaction into `ast_interaction.rs`.
- Move chat interaction functionalities to `chat_interaction.rs`.
- Structure different diff formats' logic into `diff_formats.rs`.
