+++ src/at_tools/att_patch/diff_formats.rs
@@ -1,26 +1,16 @@
-use std::collections::HashMap;
 use std::path::PathBuf;
-use std::str::FromStr;
-use std::sync::{Arc, RwLock};
-use async_trait::async_trait;
+
 use itertools::Itertools;
 use rand::distributions::Alphanumeric;
 use rand::Rng;
 use ropey::Rope;
-use serde_json::Value;
-use tokenizers::Tokenizer;
-use tracing::{info, warn};
+use tracing::warn;
 
-use crate::ast::ast_index::RequestSymbolType;
-use crate::ast::treesitter::ast_instance_structs::SymbolInformation;
 use crate::at_commands::at_commands::AtCommandsContext;
-use crate::at_commands::at_file::execute_at_file;
-use crate::at_tools::tools::Tool;
-use crate::call_validation::{ChatMessage, ChatPost, ContextEnum, DiffChunk, SamplingParameters};
+use crate::at_tools::att_patch::tool::DefaultToolPatch;
+use crate::call_validation::DiffChunk;
 use crate::diffs::apply_diff_chunks_to_text;
 use crate::files_in_workspace::{Document, read_file_from_disk};
-use crate::{cached_tokenizers, scratchpads};
-use crate::scratchpads::chat_utils_rag::count_tokens;
 
 pub struct WholeFileDiffFormat {}
 
@@ -401,306 +391,7 @@ def test_todo():
     }
 }
 
-
-const DEFAULT_MODEL_NAME: &str = "gpt-4o";
-const MAX_TOKENS: usize = 32000;
-const MAX_NEW_TOKENS: usize = 4096;
-const TEMPERATURE: f32 = 0.1;
-pub type DefaultToolPatch = UnifiedDiffFormat;
-
-
-struct PatchArguments {
-    paths: Vec<String>,
-    symbol_names: Option<Vec<String>>,
-    todo: String,
-}
-
-async fn symbols_to_signatures_context(symbols: &Vec<SymbolInformation>) -> String {
-    let mut context: String = "".to_string();
-    for s in symbols.iter() {
-        let decl_sign = match s.get_declaration_content_from_file().await {
-            Ok(sign) => sign,
-            Err(err) => {
-                warn!("Cannot get a content for symbol {:?}: {err}", s.name);
-                continue;
-            }
-        };
-        context.push_str(&format!("```\n{decl_sign}\n```\n"))
-    }
-    context
-}
-
-async fn parse_initial_arguments(
-    args: &HashMap<String, Value>,
-    ccx: &mut AtCommandsContext,
-) -> Result<PatchArguments, String> {
-    let paths = match args.get("paths") {
-        Some(Value::String(s)) => s.split(",").map(|x| x.to_string()).collect::<Vec<String>>(),
-        Some(v) => { return Err(format!("argument `path` is not a string: {:?}", v)) }
-        None => { return Err("argument `path` is not a string".to_string()) }
-    };
-    let mut corrected_paths = vec![];
-    for p in paths.into_iter() {
-        let corrected = crate::files_correction::correct_to_nearest_filename(
-            ccx.global_context.clone(),
-            &p,
-            false,
-            1,
-        ).await;
-        if corrected.is_empty() {
-            return Err(format!("Cannot find a file {p}"));
-        }
-        corrected_paths.push(corrected[0].clone());
-    }
-    let symbol_names = match args.get("symbols") {
-        Some(Value::String(s)) => Some(s.split(",").map(|x| x.to_string()).collect::<Vec<String>>()),
-        Some(v) => { return Err(format!("argument `path` is not a string: {:?}", v)) }
-        None => None
-    };
-    let todo = match args.get("todo") {
-        Some(Value::String(s)) => s.clone(),
-        Some(v) => { return Err(format!("argument `todo` is not a string: {:?}", v)) }
-        None => { "".to_string() }
-    };
-    Ok(PatchArguments {
-        paths: corrected_paths,
-        symbol_names,
-        todo,
-    })
-}
-
-async fn make_prompt(
-    args: &PatchArguments,
-    ccx: &mut AtCommandsContext,
-) -> Result<(String, Option<String>), String> {
-    let maybe_symbols = if let Some(ast_module) = ccx.global_context.read().await.ast_module.clone() {
-        let mut symbols = vec![];
-        if let Some(symbols_names) = args.symbol_names.clone() {
-            for name in symbols_names.into_iter() {
-                let res = match ast_module
-                    .read()
-                    .await
-                    .search_by_name(name, RequestSymbolType::Declaration, false, 1)
-                    .await {
-                    Ok(s) => s.search_results
-                        .get(0)
-                        .map(|x| x.symbol_declaration.clone()),
-                    Err(_) => None
-                };
-                if let Some(s) = res {
-                    symbols.push(s.clone());
-                }
-            }
-        } else {
-            for filename in args.paths.iter() {
-                if let Ok(path) = PathBuf::from_str(filename) {
-                    let doc = Document::new(&path);
-                    match ast_module
-                        .read()
-                        .await
-                        .decl_symbols_from_imports_by_file_path(&doc, 1)
-                        .await {
-                        Ok(s) => {
-                            s.search_results
-                                .iter()
-                                .map(|x| {
-                                    symbols.push(x.symbol_declaration.clone());
-                                    s.clone()
-                                })
-                                .collect::<Vec<_>>()
-                        }
-                        Err(err) => {
-                            warn!("Cannot import symbols for path {:?}: {err}", path);
-                            continue;
-                        }
-                    };
-                } else {
-                    warn!("Cannot parse path: {filename}");
-                    continue;
-                }
-            }
-        }
-        Some(symbols)
-    } else {
-        None
-    };
-
-    let extra_context = if let Some(symbols) = maybe_symbols {
-        Some(symbols_to_signatures_context(&symbols).await)
-    } else {
-        None
-    };
-    Ok((DefaultToolPatch::prompt(), extra_context))
-}
-
-async fn make_chat_history(
-    args: &PatchArguments,
-    ccx: &mut AtCommandsContext,
-    system_prompt: &String,
-    maybe_extra_context: &Option<String>,
-    tokenizer: Arc<RwLock<Tokenizer>>
-) -> Result<Vec<ChatMessage>, String> {
-    let mut tokens: usize = 0;
-    let max_tokens: usize = MAX_TOKENS - MAX_NEW_TOKENS;
-    let tokenizer_ref = tokenizer.read().unwrap().clone();
-    let task_message = format!("The task is:\n{}", args.todo).to_string();
-    let mut chat_messages = vec![
-        ChatMessage::new(
-            "system".to_string(),
-            system_prompt.to_string(),
-        )
-    ];
-    tokens += 3 + count_tokens(&tokenizer_ref, &system_prompt);
-    tokens += 3 + count_tokens(&tokenizer_ref, &task_message);
-    if tokens > max_tokens {
-        return Err(format!("Too many tokens: {tokens} > {max_tokens}"));
-    }
-
-    let has_single_file = args.paths.len() == 1;
-    for (idx, file) in args.paths.iter().enumerate() {
-        match execute_at_file(ccx, file.clone()).await {
-            Ok(res) => {
-                let message = format!("{}\n```\n{}```\n\n", res.file_name, res.file_content).to_string();
-                tokens += 3 + count_tokens(&tokenizer_ref, &message);
-                if tokens > max_tokens {
-                    let err_message = if has_single_file || idx == 0 {
-                        format!("The provided file {file} is too large for the patch tool: {tokens} > {max_tokens}")
-                    } else {
-                        format!("There are too many files provided: {tokens} > {max_tokens}. Try to call the tool for each file separately")
-                    };
-                    return Err(err_message)
-                }
-                chat_messages.push(ChatMessage::new("user".to_string(), message));
-            }
-            Err(err) => {
-                warn!("Cannot find a `{file}`: {err}");
-            }
-        }
-    }
-    if let Some(extra_context) = maybe_extra_context {
-        let message = format!("Extra context for the files:\n{}", extra_context).to_string();
-        tokens += 3 + count_tokens(&tokenizer_ref, &message);
-        if tokens > max_tokens {
-            warn!("Too many tokens for the extra context, skipping it: {tokens} > {max_tokens}");
-        } else {
-            chat_messages.push(ChatMessage::new("user".to_string(), message));
-        }
-    }
-
-    chat_messages.push(ChatMessage::new("user".to_string(), task_message));
-    info!("tokens num: {tokens}");
-    Ok(chat_messages)
-}
-
-
-async fn run_chat(
-    args: &PatchArguments,
-    ccx: &mut AtCommandsContext,
-    system_prompt: String,
-    maybe_extra_context: Option<String>,
-) -> Result<String, String> {
-    let gx = ccx.global_context.clone();
-    let mut chat_post = ChatPost {
-        messages: vec![],
-        parameters: SamplingParameters {
-            max_new_tokens: MAX_NEW_TOKENS,
-            temperature: Some(TEMPERATURE),
-            top_p: None,
-            stop: vec![],
-        },
-        model: DEFAULT_MODEL_NAME.to_string(),
-        scratchpad: "".to_string(),
-        stream: Some(false),
-        temperature: Some(TEMPERATURE),
-        max_tokens: MAX_TOKENS,
-        tools: None,
-        only_deterministic_messages: false,
-        chat_id: "".to_string(),
-    };
-    let caps = crate::global_context::try_load_caps_quickly_if_not_present(
-        gx.clone(), 0,
-    )
-        .await
-        .map_err(|e| {
-            warn!("No caps: {:?}", e);
-            "Network error communicating with the model (1)".to_string()
-        })?;
-
-    let (model_name, scratchpad_name, scratchpad_patch, n_ctx, _) = crate::http::routers::v1::chat::lookup_chat_scratchpad(
-        caps.clone(),
-        &chat_post,
-    ).await?;
-    let (client, api_key) = {
-        let cx_locked = gx.write().await;
-        (cx_locked.http_client.clone(), cx_locked.cmdline.api_key.clone())
-    };
-
-    let tokenizer = cached_tokenizers::cached_tokenizer(
-        caps.clone(), gx.clone(), model_name.clone(),
-    ).await?;
-
-    chat_post.messages = make_chat_history(
-        args, ccx, &system_prompt, &maybe_extra_context, tokenizer
-    ).await?;
-
-    let mut scratchpad = scratchpads::create_chat_scratchpad(
-        gx.clone(),
-        caps.clone(),
-        model_name.clone(),
-        &chat_post.clone(),
-        &scratchpad_name,
-        &scratchpad_patch,
-        false,
-        false,
-    ).await?;
-    let prompt = scratchpad.prompt(
-        n_ctx,
-        &mut chat_post.parameters,
-    ).await?;
-
-    let t1 = std::time::Instant::now();
-    let messages = crate::restream::scratchpad_interaction_not_stream_json(
-        gx.clone(),
-        scratchpad,
-        "chat".to_string(),
-        &prompt,
-        model_name,
-        client,
-        api_key,
-        &chat_post.parameters,
-        chat_post.only_deterministic_messages,
-    ).await.map_err(|e| {
-        warn!("Network error communicating with the (2): {:?}", e);
-        "Network error communicating with the model (2)".to_string()
-    })?;
-    info!("patch generation took {:?}ms", t1.elapsed().as_millis() as i32);
-
-    let choices_array = match messages["choices"].as_array() {
-        Some(array) => array,
-        None => return Err("Unable to get choices array from JSON".to_string()),
-    };
-
-    let choice0 = match choices_array.get(0) {
-        Some(Value::Object(o)) => o,
-        Some(v) => { return Err(format!("choice[0] is not a dict: {:?}", v)) }
-        None => { return Err("choice[0] doesn't exist".to_string()) }
-    };
-
-    let choice0_message = match choice0.get("message") {
-        Some(Value::Object(o)) => o,
-        Some(v) => { return Err(format!("choice[0].message is not a dict: {:?}", v)) }
-        None => { return Err("choice[0].message doesn't exist".to_string()) }
-    };
-
-    match choice0_message.get("content") {
-        Some(Value::String(s)) => Ok(s.clone()),
-        Some(v) => { return Err(format!("choice[0].message.content is not a string: {:?}", v)) }
-        None => { return Err("choice[0].message.content doesn't exist".to_string()) }
-    }
-}
-
-
-async fn parse_diff_from_message(
+pub async fn parse_diff_chunks_from_message(
     ccx: &mut AtCommandsContext,
     message: &String,
 ) -> Result<String, String> {
@@ -791,125 +482,5 @@ async fn parse_diff_from_message(
     match serde_json::to_string_pretty(&chunks) {
         Ok(json_chunks) => Ok(json_chunks),
         Err(err) => Err(format!("Error diff chunks serializing: {:?}", err))
-    }
-}
-
-pub struct ToolPatch {}
-#[async_trait]
-impl Tool for ToolPatch {
-    async fn execute(
-        &self,
-        ccx: &mut AtCommandsContext,
-        tool_call_id: &String,
-        args: &HashMap<String, Value>,
-    ) -> Result<Vec<ContextEnum>, String> {
-        let args = match parse_initial_arguments(args, ccx).await {
-            Ok(res) => res,
-            Err(err) => {
-                return Err(err);
-            }
-        };
-        let (system_prompt, maybe_extra_context) = match make_prompt(&args, ccx).await {
-            Ok(res) => res,
-            Err(err) => {
-                return Err(err);
-            }
-        };
-        let answer = match run_chat(&args, ccx, system_prompt, maybe_extra_context).await {
-            Ok(res) => res,
-            Err(err) => {
-                return Err(err);
-            }
-        };
-        info!("Tool patch answer: {answer}");
-        match parse_diff_from_message(ccx, &answer).await {
-            Ok(res) => {
-                info!("Tool patch diff: {:?}", res);
-                Ok(vec![(ContextEnum::ChatMessage(ChatMessage {
-                    role: "diff".to_string(),
-                    content: res,
-                    tool_calls: None,
-                    tool_call_id: tool_call_id.clone(),
-                }))])
-            }
-            Err(err) => {
-                warn!(err);
-                Err(format!("Can't make any changes: {err}"))
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_parse_diff_message() {
-        let input = r#"Let's break down the changes needed to implement the task:
-
-1. **Track the last bounding box position for each track:**
-   - Add a new dictionary `_last_bboxes` to store the last bounding box positions.
-   - Initialize this dictionary in the constructor.
-
-2. **Check for duplicates before emitting an event:**
-   - Add a method `_is_significant_movement` to check if the current bounding box has moved significantly compared to the last recorded position.
-
-3. **Emit unique events only if the bounding box has moved significantly:**
-   - Update the `_handle_abandoned_item` method to use `_is_significant_movement` before emitting an event.
-   - Update the last bounding box position after emitting an event.
-
-Here is the unified diff format for the changes:
-
-```diff
---- /home/svakhreev/projects/dssl/abandonment_prototype/abandonment/nodes/event_emitter_node.py
-+++ /home/svakhreev/projects/dssl/abandonment_prototype/abandonment/nodes/event_emitter_node.py
-@@ ... @@
-    def __init__(self, circle_radius: float, abandonment_frames: int, max_age: int, moved_frames: int,  **kwargs):
-        ProcessConsumerProducer.__init__(self, **kwargs)
-        self._circle_radius = circle_radius
-        self._abandonment_frames = abandonment_frames
-        self._max_age = max_age
-        self._moved_frames = moved_frames
-        self._abandonment_items = {}
-+        self._last_bboxes = {}  # Track last bounding box positions
-
-@@ ... @@
-    def _handle_abandoned_item(self, item: Prediction, message: Message) -> None:
-        if item.track_idx not in self._abandonment_items:
-            self._abandonment_items[item.track_idx] = AbandonmentItem(prediction=item)
-        else:
-            self._abandonment_items[item.track_idx].moved_count = 0
-            frames_diff = item.frame_idx - self._abandonment_items[item.track_idx].prediction.frame_idx
-            if frames_diff > self._abandonment_frames and not self._abandonment_items[item.track_idx].is_sent:
-+                if self._is_significant_movement(item):
-                    event = Event(prediction=self._abandonment_items[item.track_idx].prediction, event_type=EventType.DIST_2)
-                    message.data.events.append(event)
-                    self._abandonment_items[item.track_idx].is_sent = True
-+                    self._last_bboxes[item.track_idx] = item.bbox.center_xywh()[:2]  # Update last bbox position
-
-+    def _is_significant_movement(self, item: Prediction) -> bool:
-+        if item.track_idx not in self._last_bboxes:
-+            return True
-+        last_x, last_y = self._last_bboxes[item.track_idx]
-+        current_x, current_y = item.bbox.center_xywh()[:2]
-+        distance = math.sqrt((last_x - current_x) ** 2 + (last_y - current_y) ** 2)
-+        return distance > self._circle_radius / 2  # Consider significant if moved more than half the radius
-```
-
-Explanation of changes:
-1. **Constructor Update:**
-   - Added `self._last_bboxes = {}` to track the last bounding box positions. (`+`)
-
-2. **_handle_abandoned_item Method Update:**
-   - Added a check using `_is_significant_movement(item)` before emitting an event. (`+`)
-   - Updated the last bounding box position after emitting an event. (`+`)
-
-3. **New Method _is_significant_movement:**
-   - Added a new method `_is_significant_movement` to check if the current bounding box has moved significantly compared to the last recorded position. (`+`)
-
-This diff should be applied to the file `/home/svakhreev/projects/dssl/abandonment_prototype/abandonment/nodes/event_emitter_node.py`. The changes ensure that duplicate events are filtered out based on their relative position using the last bounding box in the track."#;
-        let result = DefaultToolPatch::parse_message(input).await.unwrap();
-        info!("result: {:?}", result);
     }
 }
